model: Fix concurrency issue in ensure_init()
With Ordering::Relaxed the compiler may choose to move the store before
the call to sodiumoxide::init(), and a concurrent thread may see
INITIALIZED set to true before sodiumoxide::init() finishes.

By using Acquire/Release in the load/store respectively, we guarantee
the call to init() is not ordered after the store.