model/LocationProof: Harden witness deduplication
Really make sure that all witnesses with the same witness_id
are removed. Do not assume there is only one valid signature
for each message (true for EdDSA/ed25519, not for ECDSA), nor
that ClosenessProofs are Eq (which they are not).

Move dedupe below the request consistency check: don't sanitize
bad location proofs.