Fix malicious proof convergence
We were only checking for witness-* malicious proofs for new witnesses
and prover-* proofs for new provers, so the malicious proofs view could
alternate between the minimal prover-* and witness-* proofs.

The test suite was improved to properly test convergence and all kinds
of badness in a simpler fashion.

The malicious_proofs table + trigger is now an actual view. Let the DB
figure out how to best check for malicious proofs.